Fix /api/health route: fix typo and return proper Response
- Rename heatlh -> health (directory name typo)
- Return NextResponse.json() instead of plain object (Next.js 16 requires
  proper Response/NextResponse in all branches)
- Now returns: {"success":true,"data":"OK"}
